MUIV_Application_ReturnID_Quit casted to ULONG as some apps seem to depend on it... Now closing a window of MUI apps does work on x86_64.
git-svn-id: https://svn.aros.org/svn/aros/trunk@27304 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/libs/muimaster/classes/application.h
@@ -192,7 +192,7 @@ struct MUI_InputHandlerNode
  Special Values MUIM_Application_ReturnID. Usally programm should leave
  the event loop if this is set
 **************************************************************************/
-#define MUIV_Application_ReturnID_Quit (-1)
+#define MUIV_Application_ReturnID_Quit ((ULONG)-1)
